+++ src/main/java/io/reactivex/rxjava3/internal/operators/single/SingleFlatMapIterableFlowable.java
@@ -118,7 +118,7 @@ public final class SingleFlatMapIterableFlowable<T, R> extends Flowable<R> {
 
         @Override
         public void request(long n) {
-            if (SubscriptionHelper.validate(n)) {
+            if (SubscriptionHelper.validate(((Long)(n)).longValue())) {
                 BackpressureHelper.add(requested, n);
                 drain();
             }
